# Reduce allocs in record.Marshal function