--- a/varint/varint.go
+++ b/varint/varint.go
@@ -1,11 +1,11 @@
 // Package varint implements varint and ZigZag encoding and decoding.
 package varint
 
-// https://github.com/gogo/protobuf/blob/master/proto/encode.go#L153
-func EncodeZigZag64(x int64) []byte {
+// PutZigZag64 encodes an int64 into dst using buf as buffer.
+func PutZigZag64(dst, buf []byte, x int64) []byte {
 	// use signed number to get arithmetic right shift.
-	//return EncodeVarint(uint64((x << 1) ^ uint64((int64(x) >> 63))))
-	return EncodeVarint(uint64(x<<1 ^ (x >> 63)))
+	n := PutVarint(buf, uint64(x<<1^(x>>63)))
+	return append(dst, buf[:n]...)
 }
 
 // https://github.com/gogo/protobuf/blob/master/proto/decode.go#L242
@@ -15,19 +15,17 @@ func DecodeZigZag64(buf []byte) (int64, int) {
 	return int64(x), n
 }
 
-const maxVarintBytes = 10 // maximum length of a varint
-
-// https://github.com/golang/protobuf/blob/master/proto/encode.go#L72
-func EncodeVarint(x uint64) []byte {
-	var buf [maxVarintBytes]byte
-	var n int
+// PutVarint encodes an int64 into buf and returns the number of bytes written.
+// If the buffer is too small, PutVarint will panic.
+// Based on https://github.com/golang/protobuf/blob/master/proto/encode.go#L72
+func PutVarint(buf []byte, x uint64) int {
+	n := 0
 	for n = 0; x > 127; n++ {
 		buf[n] = 0x80 | uint8(x&0x7F)
 		x >>= 7
 	}
 	buf[n] = uint8(x)
-	n++
-	return buf[0:n]
+	return n + 1
 }
 
 // https://github.com/golang/protobuf/blob/master/proto/decode.go#L57