--- a/varint/varint_test.go
+++ b/varint/varint_test.go
@@ -1,6 +1,7 @@
 package varint
 
 import (
+	"encoding/binary"
 	"math"
 	"testing"
 )
@@ -8,7 +9,9 @@ import (
 func TestZigZag64(t *testing.T) {
 	tests := []int64{0, 1, -1, math.MaxInt32, math.MinInt32, math.MaxInt64, math.MinInt64}
 	for _, tt := range tests {
-		b := EncodeZigZag64(tt)
+		buf := make([]byte, binary.MaxVarintLen64)
+		var b []byte
+		b = PutZigZag64(b, buf, tt)
 		i, _ := DecodeZigZag64(b)
 		if i != tt {
 			t.Fatal(tt, i)