The purpose of the change is to redue allocs in the hot path
by reusing buffer for varint encoding.
`EncodeZigZag64` was replaced with `PutZigZag64` extending the API
of the function so it could accept pre-allocated buffers.

The benchmark results are following:
```
benchstat  old.txt new.txt
name              old time/op    new time/op    delta
Record_Marshal-8     395ns ± 2%     268ns ±11%  -32.07%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
Record_Marshal-8      288B ± 0%      176B ± 0%  -38.89%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
Record_Marshal-8      13.0 ± 0%       6.0 ± 0%  -53.85%  (p=0.008 n=5+5)
```